requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.2.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
scikit-learn==1.6.1
scipy==1.15.1
sniffio==1.3.1
starlette==0.45.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
uvicorn==0.34.0
